Complete support for community milestones
