+++ Shared/Views/MilestoneSheet.swift
@@ -1,0 +1,17 @@
+//
+//  MilestoneSheet.swift
+//  Shuttle Tracker
+//
+//  Created by Truong Tommy on 3/20/22.
+//
+import SwiftUI
+
+struct MilestoneSheet: View {
+    
+    var body: some View {
+        NavigationView {
+            MilestoneView()
+        }
+    }
+    
+}
